main

+++ app/Http/Controllers/Api/ClientController.php
@@ -9,10 +9,23 @@ use App\Http\Requests\UpdateClientRequest;
 use App\Http\Resources\Api\ClientCollection;
 use App\Http\Resources\Api\ClientResource;
 use App\Models\Client;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\File;
 
-class ClientController extends Controller
+class ClientController extends Controller  implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('permission:view client', only: ['index', 'show']),
+            new Middleware('permission:create client', only: ['create', 'store']),
+            new Middleware('permission:update client', only: ['update', 'edit']),
+            new Middleware('permission:delete client', only: ['destroy']),
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -20,11 +33,18 @@ class ClientController extends Controller
     {
         $filter = new ClientsFilter;
         $filterItems = $filter->transform($request);
-        // $includeUser = $request->query('includeUser');
+        $includeUser = $request->query('includeUser');
+        $includeLocations = $request->query('includeLocations');
+        $includeAll = $request->query('includeAll');
         $clients = Client::where($filterItems);
-        $clients = $clients->with('user');
-
-        return new ClientCollection($clients->paginate(10000)->appends($request->query()));
+        if ($includeUser) {
+            $clients = $clients->with(['user']);
+        } elseif ($includeLocations) {
+            $clients = $clients->with(['district', 'country', 'province', 'city']);
+        } elseif ($includeAll) {
+            $clients = $clients->with(['user','district', 'country', 'province', 'city']);
+        }
+        return new ClientCollection($clients->paginate(10)->appends($request->query()));
     }
 
     /**
@@ -40,14 +60,67 @@ class ClientController extends Controller
      */
     public function store(StoreClientRequest $request)
     {
-        return new ClientResource(Client::create($request->all()));
+        $validatedData = $request->validated();
+        $client =  new ClientResource(Client::create([
+            'name' => $validatedData['full_name'],
+            'nic' => $validatedData['nic'],
+            'mobile' => $validatedData['mobile'],
+            'phone' => $validatedData['phone'],
+            'address' => $validatedData['address'],
+            'city_id' => $validatedData['city_id'],
+            'district_id' => $validatedData['district_id'],
+            'province_id' => $validatedData['province_id'],
+            'country_id' => $validatedData['country_id'],
+            'active' => $validatedData['active'],
+        ]));
+
+        if ($client) {
+            if ($request->hasFile('image')) {
+                $imageName = time() . '.' . $request->image->getClientOriginalExtension();
+                $request->image->move('img/profile/client', $imageName);
+                $validatedData['image'] = 'img/profile/client' . $imageName;
+            } else {
+                $validatedData['image'] = null;
+            }
+            $user = new ClientResource(User::create([
+                'reference_id' => $client->id,
+                'user_type_id' => 1,
+                'name' => $validatedData['name'],
+                'email' => $validatedData['email'],
+                'image' => $validatedData['image'],
+                'password' => bcrypt($validatedData['password']),
+                'last_login' => $validatedData['last_login'],
+                'active' => $validatedData['active'],
+            ]));
+            if ($user) {
+                $message = 'Client Creaeted Successfully';
+            } else {
+                $message = 'Error Occured When Creating User';
+            }
+        } else {
+            $message = 'Error Occured When Creating Client';
+        }
+
+        return response()->json([
+            'message' => $message,
+        ]);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Client $client)
+    public function show(Client $client, Request $request)
     {
+        $includeUser = $request->query('includeUser');
+        $includeLocations = $request->query('includeLocations');
+        $includeAll = $request->query('includeAll');
+        if ($includeUser) {
+            $client = $client->loadMissing(['user']);
+        } elseif ($includeLocations) {
+            $client = $client->loadMissing(['district', 'country', 'province', 'city']);
+        } elseif ($includeAll) {
+            $client = $client->loadMissing(['user', 'district', 'country', 'province', 'city']);
+        }
         return new ClientResource($client);
     }
 
@@ -64,7 +137,52 @@ class ClientController extends Controller
      */
     public function update(UpdateClientRequest $request, Client $client)
     {
-        $client->update($request->all());
+
+        $validatedData = $request->validated();
+        $client_id = $client->id;
+        $clients = $client->update([
+            'name' => $validatedData['full_name'],
+            'nic' => $validatedData['nic'],
+            'mobile' => $validatedData['mobile'],
+            'phone' => $validatedData['phone'],
+            'address' => $validatedData['address'],
+            'city_id' => $validatedData['city_id'],
+            'district_id' => $validatedData['district_id'],
+            'province_id' => $validatedData['province_id'],
+            'country_id' => $validatedData['country_id'],
+            'active' => $validatedData['active'],
+        ]);
+        if ($clients) {
+            $user = User::where('user_type_id', 1)->where('reference_id', $client_id);
+            if ($request->hasFile('image')) {
+                $imageName = time() . '.' . $request->image->getClientOriginalExtension();
+                $request->image->move('img/profile/client', $imageName);
+                $file_path = 'img/profile/client/' . $imageName;
+                $remove_old = File::delete($user->first()->image);
+                $validatedData['image'] = $file_path;
+            } else {
+                $validatedData['image'] = $user->first()->image;
+            }
+            $user->update([
+                'reference_id' => $client_id,
+                'user_type_id' => 1,
+                'name' => $validatedData['name'],
+                'email' => $validatedData['email'],
+                'image' => $validatedData['image'],
+                'active' => $validatedData['active'],
+            ]);
+            if ($user) {
+                $message = 'Client Created Successfully';
+            } else {
+                $message = 'Error Occured When Creating User';
+            }
+        } else {
+            $message = 'Error Occured When Creating Client';
+        }
+
+        return response()->json([
+            'message' => $message,
+        ]);
     }
 
     /**
@@ -72,6 +190,17 @@ class ClientController extends Controller
      */
     public function destroy(Client $client)
     {
-        $client->delete();
+        $client_id = $client->id;
+        $user = User::where('user_type_id', 1)->where('reference_id', $client_id);
+        if ($user) {
+            $user->delete();
+            $client->delete();
+            $message = 'Client And User Deleted Successfully';
+        } else {
+            $message = 'An Error Occured';
+        }
+        return response()->json([
+            'message' => $message,
+        ]);
     }
 }
